backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
python-dotenv
bcrypt